# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Permission:
    properties:
      created_at:
        type: string
      group:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Sysware Test Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sysware Test Services API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login with username and password
      parameters:
      - description: User Login Credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      summary: Login to the application
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses: {}
      summary: Register a new user
      tags:
      - Auth
  /permissions:
    get:
      consumes:
      - application/json
      description: Get all permissions with pagination and search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order by field
        in: query
        name: order_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_by
        type: string
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Search column
        in: query
        name: column
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get all permissions
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Create a new permission
      parameters:
      - description: Permission Data
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Permission
  /permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a permission
      tags:
      - Permission
    get:
      consumes:
      - application/json
      description: Get a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a permission by ID
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Update a permission by ID
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission Data
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/models.Permission'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a permission
      tags:
      - Permission
  /permissions/me:
    get:
      consumes:
      - application/json
      description: Get permissions for the authenticated user
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get my permissions
      tags:
      - Permission
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"